Refactor: 테스트 케이스 패키지 및 클래스 분리
- main 폴더의 구조와 같은 구성으로 test 패키지 구성
